requirements.txt
altgraph==0.17
APScheduler==3.6.3
autopep8==1.5.4
boto3==1.16.29
botocore==1.19.29
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Deprecated==1.2.10
docutils==0.15.2
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
orjson==3.3.0
pefile==2019.4.18
pony==0.7.13
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyGithub==1.53
pyinstaller-hooks-contrib==2020.7
PyJWT==1.7.1
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.24.0
rsa==4.5
s3transfer==0.3.3
six==1.15.0
tinyaes==1.0.1
toml==0.10.1
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
